requirements.txt
crewai
crewai-tools
python-dotenv
langchain-huggingface
